Fix iOS build

+++ Source/PLCrashAsyncSymbolication.c
@@ -89,7 +89,7 @@ plcrash_error_t plcrash_async_find_symbol (plcrash_async_macho_t *image, plcrash
     objcErr = plcrash_async_objc_find_method(image, &cache->objc_cache, pc, objc_symbol_callback, &lookup_ctx);
 
     if (machoErr != PLCRASH_ESUCCESS && objcErr != PLCRASH_ESUCCESS) {
-        PLCF_DEBUG("Could not find symbol for PC %" PRIx64 "image %p", pc, image);
+        PLCF_DEBUG("Could not find symbol for PC %" PRIx64 "image %p", (uint64_t) pc, image);
         PLCF_DEBUG("pl_async_macho_find_symbol error %d, pl_async_objc_find_method error %d", machoErr, objcErr);
         return machoErr;
     }
@@ -97,7 +97,7 @@ plcrash_error_t plcrash_async_find_symbol (plcrash_async_macho_t *image, plcrash
     /* Even if a symbol was found above, our callbacks could have errored out, in which case they would have
      * logged a debug message, not set 'found' */
     if (!lookup_ctx.found) {
-        PLCF_DEBUG("Unexpected error occured in symbol lookup callbacks for PC %" PRIx64 "image %p; returning error", pc, image);
+        PLCF_DEBUG("Unexpected error occured in symbol lookup callbacks for PC %" PRIx64 "image %p; returning error", (uint64_t) pc, image);
         return PLCRASH_EINTERNAL;
     }
